Saving changes for later

+++ ft_memcmp.c
@@ -13,10 +13,10 @@
 #include "libft.h"
 
 int ft_memcmp(void *str1, void *str2, size_t n){
-    if (ft_strlen((char*)str1) < n) return 0;
-    if (ft_strlen((char*)str2) < n) return 0;
+    if ((size_t)ft_strlen((char*)str1) < n) return 0;
+    if ((size_t)ft_strlen((char*)str2) < n) return 0;
 
-    int i = 0;
+    size_t i = 0;
     while (i < n){
         if (((char *)str1)[i] != ((char *)str2)[i])
             return 0;
